Modularized the program to follow Single Responsibility SOLID Rule. One file reads in emails via .csv. One file is the main.
